+++ react-fundamentals-course/src/helpers/formatCreationDate.js
@@ -1,0 +1,3 @@
+export const formatDate = (stringValue) => {
+    return stringValue.replace(/\//g, '.');
+}
